Move CoreDataVersionExtractor to its own file
